# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Schroedinger's REST API documentation",
    "version": "1.0.0",
    "title": "Schroedinger API",
    "contact": {
      "email": "you@schroedinger.de"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths": {
    "/user": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "create a new user account",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "recaptcha_token",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "payload",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserRegistration"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User created"
          },
          "400": {
            "description": "User could not be created"
          },
          "409": {
            "description": "Username and/or email allready exist"
          }
        }
      },
      "put": {
        "tags": [
          "user"
        ],
        "summary": "change user credentials (username, email, password)",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "payload",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ChangeUserData"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User Data was updated"
          },
          "400": {
            "description": "User Data could not be changed"
          }
        }
      },
      "delete": {
        "tags": [
          "user"
        ],
        "description": "username or email, password required",
        "summary": "user can delete his account permanently",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "payload",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DeleteUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User account was deleted"
          },
          "400": {
            "description": "User account could not be deleted"
          }
        }
      }
    },
...
    "/survey": {
      "post": {
        "tags": [
          "survey"
        ],
        "summary": "user creates a new survey",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "payload",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateSurvey"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Survey was created",
            "schema": {
              "$ref": "#/definitions/CreateSurvey"
            }
          },
          "400": {
            "description": "could not be created"
          }
        }
      }
    },
...
    "/submission": {
      "post": {
        "tags": [
          "submission"
        ],
        "summary": "user took a survey and submits it",
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "type": "string"
          },
          {
            "in": "body",
            "name": "payload",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Submission"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "submission successfull"
          },
          "400": {
            "description": "wrong parameters or id not found"
          },
          "500": {
            "description": "submission could not be handed in"
          }
        }
      },
      "get": {
        "tags": [
          "submission"
        ],
        "summary": "owner can get submissions to a survey",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "survey_id",
            "type": "string",
            "required": true
          },
          {
            "in": "query",
            "type": "string",
            "name": "page_number",
            "default": 0
          },
          {
            "in": "query",
            "type": "string",
            "name": "page_size",
            "default": 5
          }
        ],
        "responses": {
          "200": {
            "description": "found submissions to survey",
            "schema": {
              "items": {
                "$ref": "#/definitions/Submission"
              }
            }
          }
        }
      }
    },
...
    "/token": {
      "post": {
        "tags": [
          "token"
        ],
        "summary": "create and retrieve a specific number of token, in a link they can be used to provide access to secure surveys",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "payload",
            "schema": {
              "$ref": "#/definitions/TokenRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "token for links to share a survey",
            "schema": {
              "items": {
                "$ref": "#/definitions/ExampleToken"
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "token"
        ],
        "description": "used_date and submission_id are only there if the token was used. If used is not given all token (used and unused are returned)",
        "summary": "get all participation token of a survey",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "type": "string",
            "name": "survey_id",
            "required": true
          },
          {
            "in": "query",
            "type": "boolean",
            "name": "used"
          },
          {
            "in": "query",
            "type": "number",
            "name": "page_number",
            "default": 0
          },
          {
            "in": "query",
            "type": "number",
            "name": "page_size",
            "default": 3
          }
        ],
        "responses": {
          "200": {
            "description": "the participation token were fetched",
            "schema": {
              "items": {
                "$ref": "#/definitions/TokenObject"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "TokenRequest": {
      "type": "object",
      "required": [
        "survey_id",
        "amount"
      ],
      "properties": {
        "survey_id": {
          "type": "string",
          "example": "fgHd-73hg-83HZ"
        },
        "amount": {
          "type": "number",
          "example": 30
        }
      }
    },
...
    "DeleteUser": {
      "type": "object",
      "required": [
        "password"
      ],
      "properties": {
        "password": {
          "type": "string",
          "example": "superstrongpassword"
        }
      }
    },
...
    "ExampleToken": {
      "type": "object",
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "type": "string",
          "example": "exampleToken"
        }
      }
    },
    "TokenObject": {
      "type": "object",
      "required": [
        "id",
        "used",
        "used_date",
        "created",
        "survey_id",
        "submission_id"
      ],
      "properties": {
        "id": {
          "type": "string",
          "example": "jkfdsh-e9oc-wqwolje"
        },
        "used": {
          "type": "boolean",
          "example": true
        },
        "used_date": {
          "type": "string",
          "example": "25.09.2020 17:10:30.974734+02"
        },
        "created": {
          "type": "string",
          "example": "25.09.2020 17:10:30.974734+02"
        },
        "survey_id": {
          "type": "string",
          "example": "ksjd-dskjf-sdkfjh"
        },
        "submission_id": {
          "type": "string",
          "example": "sdf-dkjf-ljree-"
        }
      }
    },
    "UserRegistration": {
      "type": "object",
      "required": [
        "username",
        "password",
        "email"
      ],
      "properties": {
        "username": {
          "type": "string",
          "example": "maxmustermann"
        },
        "password": {
          "type": "string",
          "example": "astrongpassword"
        },
        "email": {
          "type": "string",
          "example": "maxmustermann@contact.de"
        }
      }
    },
...
    "ConstrainedQuestion": {
      "type": "object",
      "required": [
        "question_text",
        "position",
        "options"
      ],
      "properties": {
        "question_text": {
          "type": "string",
          "example": "Do cats have fluffy fur?"
        },
        "position": {
          "type": "number",
          "example": 2
        },
        "options": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ConstrainedOption"
          }
        }
      }
    },
...
    "ConstrainedOption": {
      "type": "object",
      "required": [
        "answer",
        "position"
      ],
      "properties": {
        "answer": {
          "type": "string",
          "example": "Very much"
        },
        "position": {
          "type": "number",
          "example": 1
        }
      }
    },
    "FreestyleQuestion": {
      "type": "object",
      "required": [
        "question_text",
        "position"
      ],
      "properties": {
        "question_text": {
          "type": "string",
          "example": "Do cats have fluffy fur?"
        },
        "position": {
          "type": "number",
          "example": 2
        }
      }
    },
...
    "AnswerdedConstrainedQuestion": {
      "type": "object",
      "required": [
        "constrained_question_id",
        "constrained_questions_option_id"
      ],
      "properties": {
        "constrained_question_id": {
          "type": "string",
          "example": "fgHd-73hg-83HZ"
        },
        "constrained_questions_option_id": {
          "type": "string",
          "example": "fgHd-73hg-83HZ"
        }
      }
    },
    "AnswerdedFreestyleQuestion": {
      "type": "object",
      "required": [
        "freestyle_question_id",
        "answer"
      ],
      "properties": {
        "freestyle_question_id": {
          "type": "string",
          "example": "fgHd-73hg-83HZ"
        },
        "answer": {
          "type": "string",
          "example": "Not that much, they scratch me often,so no"
        }
      }
    },
    "CreateSurvey": {
      "type": "object",
      "required": [
        "title",
        "description",
        "start_date",
        "end_date",
        "secured",
        "constrained_questions",
        "freestyle_questions"
      ],
      "properties": {
        "title": {
          "type": "string",
          "example": "Experience when working with Schroedinger"
        },
        "description": {
          "type": "string",
          "example": "The result of this survey is used to improve the user experience of this app"
        },
        "start_date": {
          "type": "string",
          "example": 1601478018189
        },
        "end_date": {
          "type": "string",
          "example": 1601478018189
        },
        "secured": {
          "type": "boolean",
          "example": false
        },
        "constrained_questions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ConstrainedQuestion"
          }
        },
        "freestyle_questions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/FreestyleQuestion"
          }
        }
      }
    },
...
    "Submission": {
      "type": "object",
      "required": [
        "survey_id",
        "constrained_answers",
        "freestyle_answers"
      ],
      "properties": {
        "survey_id": {
          "type": "string",
          "example": "fgHd-73hg-83HZ"
        },
        "constrained_answers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AnswerdedConstrainedQuestion"
          }
        },
        "freestyle_answers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AnswerdedFreestyleQuestion"
          }
        }
      }
    }
  }
}